clase 3/10/2026
Añadicion de constructores vacios para la impresion de el menu y la generacion de PerimetroPoligonoRegular

+++ EjecutoresEjercicios/RunnerPromedioDe3Calificaciones.cs
@@ -12,16 +12,16 @@ namespace APS_Eq_TeamAltF4_U2.EjecutoresEjercicios
     {
         public RunnerPromedioDe3Calificaciones()
         {
-            int cal1, cal2, cal3;
+            float cal1, cal2, cal3;
             Console.WriteLine("Dame la calificación 1:");
-            cal1 = int.Parse(Console.ReadLine());
+            cal1 = float.Parse(Console.ReadLine());
             Console.WriteLine("Dame la calificación 2:");
-            cal2 = int.Parse(Console.ReadLine());
+            cal2 = float.Parse(Console.ReadLine());
             Console.WriteLine("Dame la calificación 3:");
-            cal3 = int.Parse(Console.ReadLine());   
+            cal3 = float.Parse(Console.ReadLine());   
 
             Ejercicio ejercicio;
-            ejercicio = new Ejercicio_PromedioDe3Calificaciones<int>(cal1, cal2, cal3);
+            ejercicio = new Ejercicio_PromedioDe3Calificaciones<float>(cal1, cal2, cal3);
             ejercicio.Ejecutar();
         }
     }
